1. Projects_Assignment/32. Capstone Projects/2. Employee_Churn_Project_Streamlit/requirements.txt
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
base58==2.1.1
black==21.12b0
bleach==4.1.0
blinker==1.4
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
colorlover==0.3.0
cryptography==36.0.1
cufflinks==0.17.3
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.2.0
email-validator==1.1.3
entrypoints==0.4
executing==0.8.2
filelock==3.6.0
Flask==2.0.2
Flask-MySQLdb==0.2.0
Flask-WTF==1.0.0
fonttools==4.28.5
gitdb==4.0.9
GitPython==3.1.27
h11==0.13.0
htmlmin==0.1.12
huggingface-hub==0.4.0
idna==3.3
ImageHash==4.2.1
importlib-metadata==4.11.1
ipykernel==6.9.1
ipython==8.1.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.0.1
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
missingno==0.5.1
mistune==0.8.4
multimethod==1.8
mypy-extensions==0.4.3
mysqlclient==2.1.0
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.8
notebook==6.4.8
numpy==1.22.2
outcome==1.1.0
packaging==21.3
pandas==1.4.1
pandas-profiling==3.1.0
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pathspec==0.9.0
phik==0.12.2
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.4.1
plotly==5.5.0
ply==3.11
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.19.4
pure-eval==0.2.2
pyarrow==7.0.0
pycparser==2.21
pydantic==1.9.0
pydeck==0.7.1
pyforest==1.1.0
Pygments==2.11.2
Pympler==1.0.1
Pyomo==6.1.2
pyOpenSSL==21.0.0
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
pywin32==303
pywinpty==2.0.2
PyYAML==6.0
pyzmq==22.3.0
regex==2022.1.18
requests==2.27.1
sacremoses==0.0.47
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
selenium==4.1.0
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
stack-data==0.2.0
streamlit==1.6.0
tangled-up-in-unicode==0.1.0
tenacity==8.0.1
terminado==0.13.1
testpath==0.6.0
threadpoolctl==3.1.0
tokenizers==0.11.5
toml==0.10.2
tomli==1.2.3
toolz==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.16.2
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
validators==0.18.2
visions==0.7.4
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
wsproto==1.0.0
WTForms==3.0.1
xgboost==1.3.3
zipp==3.7.0
